Organize physics knowledge
Creates a utils/physics directory to centralize all physics-related functions and constants.
