+++ src/ts/backend/universe/customSystems/alphaTestis.ts
@@ -20,9 +20,9 @@ import { Tools } from "@babylonjs/core/Misc/tools";
 
 import { StarSystemCoordinates } from "@/backend/universe/starSystemCoordinates";
 
-import { celsiusToKelvin, getOrbitRadiusFromPeriod } from "@/utils/physics";
-
-import { Settings } from "@/settings";
+import { EarthSeaLevelPressure } from "@/utils/physics/constants";
+import { getOrbitRadiusFromPeriod } from "@/utils/physics/orbit";
+import { celsiusToKelvin } from "@/utils/physics/unitConversions";
 
 import { newSeededGasPlanetModel } from "../proceduralGenerators/gasPlanetModelGenerator";
 import { newSeededSpaceStationModel } from "../proceduralGenerators/orbitalFacilities/spaceStationModelGenerator";
@@ -59,7 +59,7 @@ export function getAlphaTestisSystemModel(): StarSystemModel {
 
     const ares = newSeededTelluricPlanetModel("ares", 0.3725, "Ares", [weierstrass]);
     if (ares.clouds !== null) ares.clouds.coverage = 1;
-    if (ares.atmosphere !== null) ares.atmosphere.pressure = Settings.EARTH_SEA_LEVEL_PRESSURE * 0.5;
+    if (ares.atmosphere !== null) ares.atmosphere.pressure = EarthSeaLevelPressure * 0.5;
 
     ares.orbit.semiMajorAxis = 25100 * hecate.radius;
