+++ src/ts/backend/universe/orbitalObjects/cloudsModel.ts
@@ -17,7 +17,7 @@
 
 import { Color3 } from "@babylonjs/core/Maths/math.color";
 
-import { Settings } from "@/settings";
+import { EarthSeaLevelPressure } from "@/utils/physics/constants";
 
 export type CloudsModel = {
     layerRadius: number;
@@ -44,7 +44,7 @@ export function newCloudsModel(
         specularPower: 2,
         frequency: 4,
         detailFrequency: 12,
-        coverage: 0.75 * Math.exp((-waterAmount * pressure) / Settings.EARTH_SEA_LEVEL_PRESSURE),
+        coverage: 0.75 * Math.exp((-waterAmount * pressure) / EarthSeaLevelPressure),
         sharpness: 2.5,
         color: new Color3(0.8, 0.8, 0.8),
         worleySpeed: 0.0005,
