+++ src/ts/backend/universe/proceduralGenerators/stellarObjects/blackHoleModelGenerator.ts
@@ -24,10 +24,8 @@ import { BlackHoleModel } from "@/backend/universe/orbitalObjects/stellarObjects
 
 import { GenerationSteps } from "@/utils/generationSteps";
 import { getRngFromSeed } from "@/utils/getRngFromSeed";
-import { estimateStarRadiusFromMass } from "@/utils/physics";
+import { getMassFromSchwarzschildRadius } from "@/utils/physics/blackHole";
 import { DeepReadonly } from "@/utils/types";
-
-import { Settings } from "@/settings";
 
 export function newSeededBlackHoleModel(
     id: string,
@@ -76,62 +74,4 @@ export function newSeededBlackHoleModel(
         accretionDiskRadius: blackHoleAccretionDiskRadius,
         orbit,
     };
-}
-
-/**
- * Returns the mass a black hole needs to posess a given Schwarzschild radius
- * @param radius The target radius (in meters)
- * @returns The mass needed to achieve the given radius
- */
-export function getMassFromSchwarzschildRadius(radius: number): number {
-    return (radius * Settings.C * Settings.C) / (2 * Settings.G);
-}
-
-/**
- * As the angular momentum is conserved, the black hole retains the original star's angular momentum.
- * As the original star's radius is only known approximately, the black hole's angular momentum can only be estimated.
- * The angular momentum is important in the Kerr metric to compute frame dragging.
- */
-export function estimateAngularMomentum(mass: number, rotationPeriod: number): number {
-    if (rotationPeriod === 0) return 0;
-
-    const estimatedOriginalStarRadius = estimateStarRadiusFromMass(mass);
-
-    // The inertia tensor for a sphere is a diagonal scaling matrix, we can express it as a simple number
-    const inertiaTensor = (2 / 5) * mass * estimatedOriginalStarRadius * estimatedOriginalStarRadius;
-
-    const omega = (2 * Math.PI) / rotationPeriod;
-
-    return inertiaTensor * omega;
-}
-
-/**
- * This corresponds to a=J/Mc in the Kerr metric. Physical values are between 0 and the mass of the black hole. Exceeding that range will create naked singularities. (J > M²)
- * @returns J/Mc for this black hole
- * @see https://en.wikipedia.org/wiki/Kerr_metric#Overextreme_Kerr_solutions
- */
-export function getKerrMetricA(mass: number, rotationPeriod: number): number {
-    return estimateAngularMomentum(mass, rotationPeriod) / (mass * Settings.C);
-}
-
-export function hasNakedSingularity(mass: number, rotationPeriod: number): boolean {
-    return getKerrMetricA(mass, rotationPeriod) > mass;
-}
-
-/**
- * Returns the radius of the ergosphere at a given angle theta.
- * @param mass The mass of the black hole in kilograms
- * @param rotationPeriod The rotation period of the black hole in seconds
- * @param theta The angle in radians to the black hole's rotation axis. (equator => theta = pi / 2)
- * @throws This function throws an error when the black hole is a naked singularity
- */
-export function getErgosphereRadius(mass: number, rotationPeriod: number, theta: number): number {
-    const m = (Settings.G * mass) / (Settings.C * Settings.C);
-
-    const a = getKerrMetricA(mass, rotationPeriod);
-    const cosTheta = Math.cos(theta);
-
-    if (a > m) throw new Error(`Black hole angular momentum exceeds maximum value for a Kerr black hole. a > m: ${a}`);
-
-    return m + Math.sqrt(m * m - a * a * cosTheta * cosTheta);
 }
