+++ src/ts/frontend/assets/procedural/spaceStation/habitats/helix/helixHabitat.ts
@@ -31,9 +31,8 @@ import { Transformable } from "@/frontend/universe/architecture/transformable";
 import { createHelix } from "@/utils/geometry/helixBuilder";
 import { getRngFromSeed } from "@/utils/getRngFromSeed";
 import { createEnvironmentAggregate } from "@/utils/havok";
-import { getRotationPeriodForArtificialGravity } from "@/utils/physics";
-
-import { Settings } from "@/settings";
+import { EarthG } from "@/utils/physics/constants";
+import { getRotationPeriodForArtificialGravity } from "@/utils/physics/physics";
 
 import { MetalSectionMaterial } from "../../metalSectionMaterial";
 import { HelixHabitatMaterial } from "./helixHabitatMaterial";
@@ -163,10 +162,7 @@ export class HelixHabitat implements Transformable {
     }
 
     update(cameraWorldPosition: Vector3, deltaSeconds: number) {
-        this.getTransform().rotate(
-            Axis.Y,
-            deltaSeconds / getRotationPeriodForArtificialGravity(this.radius, Settings.G_EARTH),
-        );
+        this.getTransform().rotate(Axis.Y, deltaSeconds / getRotationPeriodForArtificialGravity(this.radius, EarthG));
 
         const distanceToCamera = Vector3.Distance(cameraWorldPosition, this.getTransform().getAbsolutePosition());
