+++ src/ts/frontend/missions/nodes/actions/sightseeing/missionFlyByNode.ts
@@ -24,11 +24,11 @@ import { StarSystemCoordinates, starSystemCoordinatesEquals } from "@/backend/un
 import { StarSystemDatabase } from "@/backend/universe/starSystemDatabase";
 import { UniverseObjectId, universeObjectIdEquals } from "@/backend/universe/universeObjectId";
 
+import { lightYearsToMeters } from "@/utils/physics/unitConversions";
 import { getOrbitalObjectTypeToI18nString } from "@/utils/strings/orbitalObjectTypeToDisplay";
 import { parseDistance } from "@/utils/strings/parseToStrings";
 
 import i18n from "@/i18n";
-import { Settings } from "@/settings";
 
 import { getGoToSystemInstructions } from "../../../common";
 import { MissionContext } from "../../../missionContext";
@@ -123,7 +123,7 @@ export class MissionFlyByNode implements MissionNodeBase<MissionNodeType.FLY_BY>
     }
 
     describe(originSystemCoordinates: StarSystemCoordinates, starSystemDatabase: StarSystemDatabase): string {
-        const distance = Vector3.Distance(
+        const distanceLy = Vector3.Distance(
             starSystemDatabase.getSystemGalacticPosition(originSystemCoordinates),
             starSystemDatabase.getSystemGalacticPosition(this.targetSystemCoordinates),
         );
@@ -135,7 +135,7 @@ export class MissionFlyByNode implements MissionNodeBase<MissionNodeType.FLY_BY>
         return i18n.t("missions:sightseeing:describeFlyBy", {
             objectType: getOrbitalObjectTypeToI18nString(objectModel),
             systemName: systemModel.name,
-            distance: distance > 0 ? parseDistance(distance * Settings.LIGHT_YEAR) : i18n.t("missions:common:here"),
+            distance: distanceLy > 0 ? parseDistance(lightYearsToMeters(distanceLy)) : i18n.t("missions:common:here"),
         });
     }
