+++ src/ts/frontend/universe/orbitalFacility/spaceElevator.ts
@@ -45,7 +45,7 @@ import { getEdibleEnergyPerHaPerDay } from "@/utils/agriculture";
 import { getRngFromSeed } from "@/utils/getRngFromSeed";
 import { isSizeOnScreenEnough } from "@/utils/isObjectVisibleOnScreen";
 import { clamp, remap, triangleWave } from "@/utils/math";
-import { getSphereRadiatedEnergyFlux } from "@/utils/physics";
+import { getSphereRadiatedEnergyFlux } from "@/utils/physics/thermodynamics";
 import { wheelOfFortune } from "@/utils/random";
 import { getSolarPanelSurfaceFromEnergyRequirement } from "@/utils/solarPanels";
 import { getOrbitalObjectTypeToI18nString } from "@/utils/strings/orbitalObjectTypeToDisplay";
