+++ src/ts/frontend/universe/stellarObjects/blackHole/blackHoleUniforms.ts
@@ -20,10 +20,10 @@ import { CubeTexture } from "@babylonjs/core/Materials/Textures/cubeTexture";
 import { TransformNode } from "@babylonjs/core/Meshes/transformNode";
 
 import { BlackHoleModel } from "@/backend/universe/orbitalObjects/stellarObjects/blackHoleModel";
-import { getKerrMetricA } from "@/backend/universe/proceduralGenerators/stellarObjects/blackHoleModelGenerator";
 
 import { getForwardDirection } from "@/frontend/uberCore/transforms/basicTransform";
 
+import { getKerrMetricA } from "@/utils/physics/blackHole";
 import { DeepReadonly } from "@/utils/types";
 
 export const BlackHoleUniformNames = {
