+++ src/ts/utils/physics/constants.ts
@@ -1,0 +1,72 @@
+//  This file is part of Cosmos Journeyer
+//
+//  Copyright (C) 2024 Barthélemy Paléologue <barth.paleologue@cosmosjourneyer.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+/**
+ * The number of meters in a light year.
+ */
+export const LightYearInMeters = 3e8 * 60 * 60 * 24 * 365.25;
+
+/**
+ * Speed of light in meters per second.
+ * @see https://en.wikipedia.org/wiki/Speed_of_light
+ */
+export const C = 299_792_458;
+
+/**
+ * The gravitational constant in m^3 kg^-1 s^-2.
+ */
+export const G = 6.6743e-11;
+
+/**
+ * The astronomical unit in meters.
+ */
+export const AU = 150e9;
+
+/**
+ * The mass of the Earth in kilograms.
+ */
+export const EarthMass = 5.972e24;
+
+/**
+ * The mass of the Moon in kilograms.
+ */
+export const MoonMass = 7.348e22;
+
+/**
+ * The mass of Jupiter in kilograms.
+ */
+export const JupiterMass = 1.898e27;
+
+/**
+ * The mass of the sun in kilograms.
+ */
+export const SolarMass = 1.989e30;
+
+/**
+ * The radius of the sun in meters.
+ */
+export const SolarRadius = 696340e3;
+
+/**
+ * The gravitational acceleration in m/s^2.
+ */
+export const EarthG = 9.81;
+
+/**
+ * The sea level pressure on Earth in Pascals.
+ */
+export const EarthSeaLevelPressure = 101_325;
