+++ src/ts/utils/physics/orbit.spec.ts
@@ -1,0 +1,40 @@
+//  This file is part of Cosmos Journeyer
+//
+//  Copyright (C) 2024 Barthélemy Paléologue <barth.paleologue@cosmosjourneyer.com>
+//
+//  This program is free software: you can redistribute it and/or modify
+//  it under the terms of the GNU Affero General Public License as published by
+//  the Free Software Foundation, either version 3 of the License, or
+//  (at your option) any later version.
+//
+//  This program is distributed in the hope that it will be useful,
+//  but WITHOUT ANY WARRANTY; without even the implied warranty of
+//  MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+//  GNU Affero General Public License for more details.
+//
+//  You should have received a copy of the GNU Affero General Public License
+//  along with this program.  If not, see <https://www.gnu.org/licenses/>.
+
+import { describe, expect, test } from "vitest";
+
+import { getOrbitRadiusFromPeriod } from "./orbit";
+
+describe("getOrbitRadiusFromPeriod", () => {
+    test("earth orbit", () => {
+        const sunMass = 1.989e30; // in kg
+        const earthPeriod = 365.25 * 24 * 60 * 60; // in seconds
+        const earthOrbitRadius = getOrbitRadiusFromPeriod(earthPeriod, sunMass);
+
+        expect(earthOrbitRadius).toBeGreaterThan(1.49e11 - 1e9);
+        expect(earthOrbitRadius).toBeLessThan(1.49e11 + 1e9);
+    });
+
+    test("moon orbit", () => {
+        const earthMass = 5.972e24;
+        const moonPeriod = 27.3 * 24 * 60 * 60; // in seconds
+        const moonOrbitRadius = getOrbitRadiusFromPeriod(moonPeriod, earthMass);
+
+        expect(moonOrbitRadius).toBeGreaterThan(3.84e8 - 2e6);
+        expect(moonOrbitRadius).toBeLessThan(3.84e8 + 2e6);
+    });
+});
